Url sanitize (#1873)
* Adds basic url sanitization.
- Clamps the range of the shape, move, and turn blocks to be in the range supported by the drop down.
- Adds a regex to make sure colors are valid 6 or 3 digit hex values.
- Makes loop blocks max out at 99.

In all of these cases if the value is out of range or invalid, it sets the value back to the default the block
has in the flyout. Also updated the block definitions so they always show an image even
if we don't have one for the value loaded from the url.