Enhance weather and tweak snowball intro
This change adds a "directional strength" property to santa-weather. The
default value of this property is 0, which retains the behavior of the
element prior to this change. When the value is negative, it adds a
constant "wind force" towards the left of the viewport. When the value
is positive, it adds a constant force towards the right of the viewport.

In addition to adding the directional strength, the snow placement has
been enhanced a bit. Snow can now spawn at the sides of the viewport in
addition to the top. Depending on the directional strength, snow will be
more likely to spawn at one side or the other.

Finally, the first round of snow now spawns within the viewport so that
a scene using santa-weather will have the appearance of being in the
snowfall when the viewer arrives.

The snowball scene has been tweaked a little. The sleigh now does a
nyan-cat style fly-through of the scene, and the directional force of
the weather has been tweaked to give the sleigh the illusion of some
velocity. Also, the tutorial is now properly shut off (it was causing a
50% frame time hit by running in the background at the same time as the
game).