Adds pose-boogie skeleton & debug pose feed.
Skeleton is pretty basic but includes html, scss & js. I've c/p'd in bits of code from the posenet demo[1], but only bits we need, and mapped to a structure I think we want.

The code is largely not mine but happy to take feedback on it still. I'd also like your opinion on the structure before I add more stuff.

TIA ✌️

[1]: https://github.com/tensorflow/tfjs-models/blob/master/posenet/demos/camera.html