Adds simple upper body animation to poseboogie
The physics are far from perfect, the legs are missing, and the arms need elbows. There are lots of TODOs in the pose mapping logic (e.g. smoothing movement, detecting out-of-frame parts, ignoring low-confidence parts and more) but this is an OK start that sometimes looks OK if you get everything in the video frame.