Fixes shoulders onto the body's defined pivots.
This means arms are no longer absolutely positioned, but rather relative
to the torso. This is mostly fine, unless the torso is not detected, and
there's no other grounding parts (e.g. feet, or more than any 1 part),
in which case jittery detections can send the body spinning around the
one fixed part.