I added new physically based present throwing
We don't have a physics engine in the game yet (and with our current scoring system, I don't think we want one?).

Instead I've moved from a cubic b-spline to a naive computation of initial y velocity given the distance to the target and how fast we want the present to travel in the x/z plane. This is hard to tune for our test level, but I think when we have better targets in place this will feel more natural?