Make the presents spawn from the elf
Now the presents are actually thrown from the right place. The place is
just an empty object in the scene with a specific name -- I think a
lightweight 'tagging' system like that is probably going to be the best
for this game :)

Patrick's code that lets us set the present's parent was very helpful!