Update railroad ride game with more detailed scene (#236)
* Remove vestigial mocha-headless-server file to fix npm deps

This was previously used for tests when Santa Tracker had a separate API, but now it's no longer used. It depends on a specific version of the the dhost library, which messes with npm dependency resolution

Both are just used for development so don't impact prod builds

* Update some dependencies (npm audit fix)

* Add new railroad ride scene

There's a bit of a jump up in filesize, but hopefully it's ok. Definitely could be optimized more (not such in the scene is instanced, I believe).